requirements.txt
flask==2.0.1
numpy==1.23.0
opencv-python==4.6.0.66
Pillow==9.2.0
torch==2.0.1
torchvision==0.15.2
scipy==1.10.1
matplotlib==3.7.1
scikit-image==0.19.3
tqdm==4.66.1
requests==2.31.0
